+++ Light.h
@@ -17,9 +17,9 @@ public:
     void setColor(Vec3 color) { m_color = color; }
     Vec3& getCenter() { return m_center; }
     Vec3& getColor() { return m_color; }
-    float getPower() { return m_color.Power(); }
+    double getPower() { return m_color.Power(); }
     virtual Ray GetEmitRay() = 0;
-    virtual InterResult intersection(Ray &ray, float &dist) = 0;
+    virtual InterResult intersection(Ray &ray, double &dist) = 0;
 protected:
     Vec3 m_center, m_color;
 };
@@ -29,7 +29,7 @@ class SpotLight: public Light
 public:
     SpotLight(Vec3 center) : Light(center) {}
     virtual Ray GetEmitRay();
-    virtual InterResult intersection(Ray &ray, float &dist) {
+    virtual InterResult intersection(Ray &ray, double &dist) {
         return MISS;
     }
 };
@@ -39,7 +39,7 @@ public:
 class SquareLight: public Light
 {
 public:
-    SquareLight(float y, float x0, float x1, float z0, float z1) : Light() {
+    SquareLight(double y, double x0, double x1, double z0, double z1) : Light() {
         m_norm.set(0, 1.0, 0);
         m_x0 = x0; m_x1 = x1; m_z0 = z0; m_z1 = z1;
         m_y = y;
@@ -49,11 +49,11 @@ public:
     void setBorder(int x0, int x1, int z0, int z1) {
         m_x0 = x0; m_x1 = x1; m_z0 = z0; m_z1 = z1;
     }
-    void sety(float y) { m_y = y; }
-    virtual InterResult intersection(Ray &ray, float &dist);
+    void sety(double y) { m_y = y; }
+    virtual InterResult intersection(Ray &ray, double &dist);
 private:
     Vec3 m_norm;
-    float m_x0, m_x1, m_z0, m_z1, m_y;
+    double m_x0, m_x1, m_z0, m_z1, m_y;
 };
