+++ Object/Plane.cpp
@@ -4,15 +4,15 @@
 
 #include "Plane.h"
 
-InterResult Plane::intersection(Ray &ray, float &dist)
+InterResult Plane::intersection(Ray &ray, double &dist)
 {
     InterResult res = MISS;
-    float t = m_normal.Dot(ray.GetDirection());
+    double t = m_normal.Dot(ray.GetDirection());
     //printf("t: %.6f\n", t);
     if (t == 0.0) return MISS;
-    float v = - m_normal.Dot(ray.GetOrigin()) - m_d;
+    double v = - m_normal.Dot(ray.GetOrigin()) - m_d;
     //printf("v: %.6f\n", v);
-    float dist_tmp = v / t;
+    double dist_tmp = v / t;
     if (dist_tmp <= 0) return MISS;
     else if (dist_tmp >= dist) return MISS;
     else {
