+++ Object/Sphere.cpp
@@ -4,16 +4,16 @@
 
 #include "Sphere.h"
 
-InterResult Sphere::intersection(Ray &ray, float &dist)
+InterResult Sphere::intersection(Ray &ray, double &dist)
 {
     InterResult result = MISS;
     Vec3 v = m_origin - ray.GetOrigin();
-    float a = v.Dot(ray.GetDirection());
-    float b = m_radius * m_radius - v.SqrLength() + a * a;
+    double a = v.Dot(ray.GetDirection());
+    double b = m_radius * m_radius - v.SqrLength() + a * a;
     if (b > 0) {
         b = sqrt(b);
         if (a > b) {
-            float dist_tmp = a - b;
+            double dist_tmp = a - b;
             if (dist_tmp >= dist) result = MISS;
             else {
                 result = HIT;
@@ -21,7 +21,7 @@ InterResult Sphere::intersection(Ray &ray, float &dist)
             }
         }
         else if (b > a && a > (-b)) {
-            float dist_tmp = b + a;
+            double dist_tmp = b + a;
             if (dist_tmp >= dist) result = MISS;
             else {
                 result = INSIDE_HIT;
