+++ RayTrace.h
@@ -8,10 +8,10 @@
 #include "Scene.h"
 #include "Camera.h"
 
-const float MIN_WEIGHT = 0.0001;
+const double MIN_WEIGHT = 0.0001;
 const int MAX_DEPTH = 5;
 
-void RayTrace(Ray ray, int depth, float weight, Vec3 &color, Scene &scene);
+void RayTrace(Ray ray, int depth, double weight, Vec3 &color, Scene &scene);
 void DrawWithRayTrace(Camera &camera, Scene &scene);
 
 #endif //LIGHT_RAYTRACE_H
